move common stuff to kademlia.go, add XOR distance
